Move user finder sub classes to their own files
I'm having a hard time following what's going on with these
classes and I don't think 4 class definititions + inheritance in the
same file is helping make that easier.
