+++ lib/mediators/messages/user_finder.rb
@@ -59,97 +59,4 @@ module Mediators::Messages
       }
     end
   end
-
-  class EmailUserFinder < UserFinder
-    private
-    def get_notifiables
-      self.users_details = Recipient.find_active_by_app_id(app_id: target_id).map do |r|
-        extract_user(:self, { "email" => r.email, "id" => r.id })
-      end
-    end
-
-    def update_or_create_user(hid:, email:)
-      Recipient[hid]
-    end
-  end
-
-  class UserUserFinder < UserFinder
-    private
-    def get_notifiables
-      user_response = heroku_client.account_info(user_uuid: target_id)
-
-      id = user_response.fetch('id')
-
-      if id != target_id
-        raise "Mismatching ids, asked for #{target_id}, got #{id}"
-      end
-
-      if user_response.fetch('last_login')
-        self.users_details = [ extract_user(:self, user_response) ]
-      else
-        self.users_details = [ ]
-      end
-    rescue Telex::HerokuClient::NotFound
-      self.users_details = [ ]
-      Telex::Sample.count "user_not_found"
-    end
-  end
-
-  class AppUserFinder < UserFinder
-    private
-    def get_notifiables
-      if app_info.nil?
-        self.users_details = [ ]
-        return
-      end
-
-      self.users_details = (owners + collabs).uniq {|u| u[:email] }.select do |user|
-        # This filters out users who have never logged in
-        UserUserFinder.run(target_id: user[:hid]).present?
-      end
-    end
-
-    def app_info
-      @app_info ||= heroku_client.app_info(target_id)
-    rescue Telex::HerokuClient::NotFound
-      Pliny.log(missing_app: true, app_id: target_id)
-      Telex::Sample.count "app_not_found"
-      nil
-    end
-
-    def owners
-      owner = extract_user(:owner, app_info.fetch('owner'))
-      if owner[:email].end_with?('@herokumanager.com')
-        team_users(owner[:email])
-      else
-        [ owner ]
-      end
-    end
-
-    def team_users(owner_email)
-      team_members_response = heroku_client.team_members(owner_email.split('@').first)
-      team_admins = team_members_response.select { |member| member.fetch('role') == 'admin' }
-      team_admins.map do |admin|
-        extract_user(:owner, admin.fetch('user'))
-      end
-    rescue Telex::HerokuClient::NotFound
-      # Organization is missing
-      Pliny.log(missing_team: true, team: owner_email)
-      Telex::Sample.count "team_not_found"
-      []
-    end
-
-    def collabs
-      collab_response = heroku_client.app_collaborators(target_id)
-      collab_response.map do |row|
-        extract_user(:collaborator, row.fetch('user'))
-      end.compact
-    rescue Telex::HerokuClient::NotFound
-      # Between the time we looked up the app in app_info and now, the app
-      # has been deleted.
-      # Don't bother sampling since this is only a fluke.
-      Pliny.log(missing_app_on_collab_lookup: true, app_id: target_id)
-      []
-    end
-  end
 end
